Does not work yet: Added base abstract class for Aspnet tests, both normal and composite.
